fin quete 24 route

+++ src/Controller/CommentController.php
@@ -7,6 +7,7 @@ use App\Entity\Episode;
 use App\Entity\User;
 use App\Form\CommentType;
 use App\Repository\CommentRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -23,6 +24,7 @@ class CommentController extends AbstractController
         ]);
     }
 
+    #[IsGranted('ROLE_CONTRIBUTOR')]
     #[Route('/new/{episode}/{user}', name: 'app_comment_new', methods: ['GET', 'POST'])]
     public function new(Request $request, Episode $episode, User $user, CommentRepository $commentRepository): Response
     {
@@ -56,9 +58,16 @@ class CommentController extends AbstractController
         ]);
     }
 
+    #[IsGranted('ROLE_CONTRIBUTOR')]
     #[Route('/{id}/edit', name: 'app_comment_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Comment $comment, CommentRepository $commentRepository): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if ($this->getUser() !== $comment->getAuthor()) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw $this->createAccessDeniedException('Seul l\'auteur de ce commentaire peut le modifier !');
+        }
+
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
@@ -74,9 +83,16 @@ class CommentController extends AbstractController
         ]);
     }
 
+    #[IsGranted('ROLE_CONTRIBUTOR')]
     #[Route('/{id}', name: 'app_comment_delete', methods: ['POST'])]
     public function delete(Request $request, Comment $comment, CommentRepository $commentRepository): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if ($this->getUser() !== $comment->getAuthor()) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw $this->createAccessDeniedException('Seul l\'auteur de ce commentaire peut le supprimer !');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $comment->getId(), $request->request->get('_token'))) {
             $commentRepository->remove($comment, true);
         }
